% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafficaccidents.R
\docType{data}
\name{calendar}
\alias{calendar}
\title{Calendar}
\format{Data frame with columns
\describe{
\item{date}{Date}
\item{date_POSIXct}{Date and time (0h)}
\item{year}{Year}
\item{month}{Month}
\item{day}{Day}
\item{wday}{Week day}
\item{year_day}{Year day}
\item{working_day}{"Working day" or "Weekend or holliday"}
}}
\usage{
calendar
}
\description{
Calendar from 2005 - 2016
}
\examples{
  head(calendar)
}
\keyword{datasets}
