% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafficaccidents.R
\docType{data}
\name{event}
\alias{event}
\title{Traffic accidents}
\format{Data frame with columns
\describe{
\item{src_file}{A source file (year)}
\item{event_id}{Event id - unique within the src_file}
\item{injury}{Injury category}
\item{district}{Country district}
\item{date}{Date of the event}
\item{hour}{Hour of the event}
\item{in_city}{Is in the city}
\item{road_type}{Road category}
\item{road_mark}{Road id}
\item{location}{City or street}
\item{road_section_mark}{Section code}
\item{road_section}{Section}
\item{pos_number}{Address}
\item{location_type}{}
\item{cause}{Main cause type}
\item{event_type}{Type of accident}
\item{weather}{Weather}
\item{traffic}{Traffic}
\item{surface_conditions}{Surface condition}
\item{surface}{Type of surface}
\item{pos_x}{Position D48}
\item{pos_y}{Position D48}
\item{lon}{Position WGS84}
\item{lat}{Position WGS84}
}}
\source{
http://www.policija.si/index.php/statistika/prometna-varnost

  http://www.policija.si/baza/pnYYYY.zip
   (where YYYY in 05 to 2016)
}
\usage{
event
}
\description{
Accident events examined by police from 2005 to 2016
}
\examples{
  head(event)
}
\keyword{datasets}
