% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafficaccidents.R
\docType{data}
\name{party}
\alias{party}
\title{Parties}
\format{Data frame with columns
\describe{
\item{src_file}{Source file (year)}
\item{event_id}{Event id from event - use with src_file}
\item{accident_role}{At fault or just involved}
\item{age}{Party age}
\item{gender}{Gender}
\item{country}{Country name}
\item{p_injury}{Party injury severity}
\item{traffic_role}{Traffic role - e.g. Car driver, Truck driver ...}
\item{seat_belt_or_helmet}{}
\item{experience_y}{Experience - years}
\item{experience_m}{Experience - months}
\item{alcotest}{Result of alcotest}
\item{alco_exam}{Result of alco exam}
\item{country_iso}{2-letter country id}
\item{p_district}{Party district}
}}
\source{
http://www.policija.si/index.php/statistika/prometna-varnost

  http://www.policija.si/baza/pnYYYY.zip, PNL-OSEBE-yyyy.TXT
   (where YYYY in 05 to 2016)
}
\usage{
party
}
\description{
Parties involved in traffic accidents
}
\examples{
  head(party)
}
\keyword{datasets}
